Cleaner support for JupyterHub + Binder
JupyterHub sets a JUPYTERHUB_SERVICE_PREFIX environment variable with
the base URL of the currently running user server. This takes into account
various factors, such as the base URL of the JupyterHub itself, named
servers (if the user has multiple servers running), etc. Binder also
sets the same environment variable, since it uses JupyterHub behind the
scenes.

jupyter-server-proxy proxies arbitrary HTTP requests sent to
$JUPYTERHUB_SERVICE_PREFIX/proxy/<port><path> to localhost:<port><path>
on the server.

This transformer assumes you are running on a JupyterHub (or Binder),
and constructs the appropriate URL for vega to reach the altair server.

You can optionally pass in `urlpath` to override the default.

This change is backwards compatible, since passing urlpath will
still work. The default is just a lot more robust.